move generate and fetch models to related folder
