+++ src/teehr/fetching/models/__init__.py
@@ -1,0 +1,1 @@
+"""Fetching model definitions."""
