+++ src/teehr/fetching/models/nwm20_grid.py
@@ -1,11 +1,11 @@
-"""Module describing NWM v3.0 grid configuration variables."""
+"""Module describing NWM v2.0 grid configuration variables."""
 from teehr.models.str_enum import StrEnum
 
 from typing import Optional
 
 from pydantic import BaseModel
 
-from teehr.models.fetching.nwm22_grid import (
+from teehr.fetching.models.nwm22_grid import (
     Analysis,
     ShortRange,
     MediumRange,
@@ -14,24 +14,16 @@ from teehr.models.fetching.nwm22_grid import (
 )
 
 
-# CONFIGURATIONS ENUM: All possible configuration
+# CONFIGURATIONS ENUM
 class ConfigurationsEnum(StrEnum):
     """ConfigurationsEnum."""
+
     analysis_assim = "analysis_assim"
-    analysis_assim_no_da = "analysis_assim_no_da"
     analysis_assim_extend = "analysis_assim_extend"
-    analysis_assim_extend_no_da = "analysis_assim_extend_no_da"
     analysis_assim_long = "analysis_assim_long"
-    analysis_assim_long_no_da = "analysis_assim_long_no_da"
     analysis_assim_hawaii = "analysis_assim_hawaii"
-    analysis_assim_hawaii_no_da = "analysis_assim_hawaii_no_da"
-    analysis_assim_puertorico = "analysis_assim_puertorico"
-    analysis_assim_puertorico_no_da = "analysis_assim_puertorico_no_da"
     short_range = "short_range"
     short_range_hawaii = "short_range_hawaii"
-    short_range_puertorico = "short_range_puertorico"
-    short_range_hawaii_no_da = "short_range_hawaii_no_da"
-    short_range_puertorico_no_da = "short_range_puertorico_no_da"
     medium_range_mem1 = "medium_range_mem1"
     medium_range_mem2 = "medium_range_mem2"
     medium_range_mem3 = "medium_range_mem3"
@@ -44,39 +36,24 @@ class ConfigurationsEnum(StrEnum):
     long_range_mem3 = "long_range_mem3"
     long_range_mem4 = "long_range_mem4"
     forcing_medium_range = "forcing_medium_range"
-    forcing_medium_range_blend = "forcing_medium_range_blend"
-    forcing_medium_range_alaska = "forcing_medium_range_alaska"
-    forcing_medium_range_blend_alaska = "forcing_medium_range_blend_alaska"
     forcing_short_range = "forcing_short_range"
     forcing_short_range_hawaii = "forcing_short_range_hawaii"
-    forcing_short_range_puertorico = "forcing_short_range_puertorico"
-    forcing_short_range_alaska = "forcing_short_range_alaska"
     forcing_analysis_assim = "forcing_analysis_assim"
     forcing_analysis_assim_extend = "forcing_analysis_assim_extend"
     forcing_analysis_assim_hawaii = "forcing_analysis_assim_hawaii"
-    forcing_analysis_assim_puertorico = "forcing_analysis_assim_puertorico"
-    forcing_analysis_assim_alaska = "forcing_analysis_assim_alaska"
 
 
 # CONFIGURATION MODEL
 class GridConfigurationModel(BaseModel):
-    """NWM v3.0 GridConfigurationModel."""
+    """NWM v2.0 GridConfigurationModel."""
+
     configuration: ConfigurationsEnum
     analysis_assim: Optional[Analysis] = None
-    analysis_assim_no_da: Optional[Analysis] = None
     analysis_assim_extend: Optional[Analysis] = None
-    analysis_assim_extend_no_da: Optional[Analysis] = None
     analysis_assim_long: Optional[Analysis] = None
-    analysis_assim_long_no_da: Optional[Analysis] = None
     analysis_assim_hawaii: Optional[Analysis] = None
-    analysis_assim_hawaii_no_da: Optional[Analysis] = None
-    analysis_assim_puertorico: Optional[Analysis] = None
-    analysis_assim_puertorico_no_da: Optional[Analysis] = None
     short_range: Optional[ShortRange] = None
     short_range_hawaii: Optional[ShortRange] = None
-    short_range_puertorico: Optional[ShortRange] = None
-    short_range_hawaii_no_da: Optional[ShortRange] = None
-    short_range_puertorico_no_da: Optional[ShortRange] = None
     medium_range_mem1: Optional[MediumRange] = None
     medium_range_mem2: Optional[MediumRange] = None
     medium_range_mem3: Optional[MediumRange] = None
@@ -84,29 +61,21 @@ class GridConfigurationModel(BaseModel):
     medium_range_mem5: Optional[MediumRange] = None
     medium_range_mem6: Optional[MediumRange] = None
     medium_range_mem7: Optional[MediumRange] = None
-    medium_range_no_da: Optional[MediumRange] = None
     long_range_mem1: Optional[LongRange] = None
     long_range_mem2: Optional[LongRange] = None
     long_range_mem3: Optional[LongRange] = None
     long_range_mem4: Optional[LongRange] = None
     forcing_medium_range: Optional[Forcing] = None
-    forcing_medium_range_blend: Optional[Forcing] = None
-    forcing_medium_range_alaska: Optional[Forcing] = None
-    forcing_medium_range_blend_alaska: Optional[Forcing] = None
     forcing_short_range: Optional[Forcing] = None
     forcing_short_range_hawaii: Optional[Forcing] = None
-    forcing_short_range_puertorico: Optional[Forcing] = None
-    forcing_short_range_alaska: Optional[Forcing] = None
     forcing_analysis_assim: Optional[Forcing] = None
     forcing_analysis_assim_extend: Optional[Forcing] = None
     forcing_analysis_assim_hawaii: Optional[Forcing] = None
-    forcing_analysis_assim_puertorico: Optional[Forcing] = None
-    forcing_analysis_assim_alaska: Optional[Forcing] = None
 
 
 if __name__ == "__main__":
     # So for example:
-    configuration = "forcing_medium_range_blend_alaska"
+    configuration = "forcing_medium_range"
     output_type = "forcing"
     variable_name = "RAINRATE"
 
@@ -118,19 +87,9 @@ if __name__ == "__main__":
             output_type: variable_name,
         },
     }
-
-    # import sys
-    # Check input parameters
-    # try:
     cm = GridConfigurationModel.model_validate(vars)
-    # except ValidationError as e:
-    #     print(e.errors()[0]["msg"])
-    #     sys.exit()
 
     config = cm.configuration.name
     forecast_obj = getattr(cm, config)
     out_type = forecast_obj.output_type.name
-    # out_type = getattr(cm, config).output_type.name
     var_name = getattr(forecast_obj, out_type).name
-
-    pass
