+++ src/teehr/fetching/models/nwm20_point.py
@@ -6,7 +6,7 @@ from typing import Optional
 from pydantic import BaseModel
 
 
-from teehr.models.fetching.nwm22_point import (
+from teehr.fetching.models.nwm22_point import (
     ShortAndAnalysis,
     MediumRangeMem1,
     MediumRangeMem2,
