+++ src/teehr/fetching/models/utils.py
