+++ src/teehr/generate/models/__init__.py
@@ -1,0 +1,1 @@
+"""Generate model definitions."""
