+++ src/teehr/generate/models/base.py
