+++ src/teehr/generate/utils.py
@@ -9,7 +9,7 @@ from pyspark.sql import functions as F
 import pyspark.sql as ps
 from pyspark.sql import Window
 
-from teehr.models.generate.base import (
+from teehr.generate.models.base import (
     NormalsResolutionEnum
 )
 from teehr.calculated_fields.models.row_level import (
