--- src/teehr/models/fetching/__init__.py
@@ -1,1 +1,0 @@
-"""Data fetching models for TEEHR."""
